Adding examples of use the FilteredTable
